# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /api/v1.0
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  Bearer:
    description: For accessing the API a valid JWT token must be passed in all the queries in the 'Authorization' header.
    type: apiKey
    name: Authorization
    in: header  
paths:
...
  /restaurant:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get the restaurants list
      operationId: getRestaurantAll
      # define the type of response for Success "200" and Error          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRestaurantsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: add a new restaurant to the list
      operationId: saveRestaurant
      # restaurant info to be stored
      parameters:
        - name: title
          description: Restaurant properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Restaurant"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
  /restaurant/{id}:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant
      operationId: getRestaurantOne
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRestaurantResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
    put:
      description: update a restaurant
      operationId: updateRestaurant
      # define the parameters
      parameters:
        - name: id
          description: Restaurant id
          type: string
          in: path
          required: true
        - name: title
          description: Restaurant properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Restaurant"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
    delete:
      description: delete a restaurant
      operationId: delRestaurant
      # define the parameters
      parameters:
        - name: id
          description: Restaurant id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /restaurant/{rid}/customers:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant customers
      operationId: getRestaurantCustomer
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetCustomersListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /restaurant/{rid}/customers/{id}:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant customer
      operationId: getRestaurantCustomer
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true
        - name: id
          type: string
          in: path
          required: true          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetCustomerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /restaurant/{rid}/waits:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant waits
      operationId: getRestaurantWaits
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetWaitsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /restaurant/{rid}/waits/{id}:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant waits
      operationId: getRestaurantWaits
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true
        - name: id
          type: string
          in: path
          required: true          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetWaitsResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /restaurant/{rid}/dinetables:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant dinetables
      operationId: getRestaurantDineTable
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetDineTablesListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /restaurant/{rid}/dinetables/{id}:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant dinetable
      operationId: getRestaurantDineTable
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true
        - name: id
          type: string
          in: path
          required: true          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetDineTableResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /restaurant/{rid}/dishgroups:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant dishgroups
      operationId: getRestaurantDishgroup
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetDishgroupsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /restaurant/{rid}/dishgroups/{id}:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant dishgroup
      operationId: getRestaurantDishgroup
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true
        - name: id
          type: string
          in: path
          required: true          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetDishgroupResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /restaurant/{rid}/dishes:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant dishes
      operationId: getRestaurantDish
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetDishesListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /restaurant/{rid}/dishes/{id}:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant dish
      operationId: getRestaurantDish
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true
        - name: id
          type: string
          in: path
          required: true          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetDishResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /restaurant/{rid}/orders:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant order
      operationId: getRestaurantOrder
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetOrdersListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /restaurant/{rid}/orders/{id}:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant order
      operationId: getRestaurantOrder
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true
        - name: id
          type: string
          in: path
          required: true          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetOrderItemsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /restaurant/{rid}/receipts:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant order
      operationId: getRestaurantReceipt
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetReceiptsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /restaurant/{rid}/receipts/{id}:
    # our controller name
    x-swagger-router-controller: restaurant
    get:
      description: get a restaurant order
      operationId: getRestaurantReceipt
      # define the type of response for Success "200" and Error
      parameters:
        - name: rid
          type: string
          in: path
          required: true
        - name: id
          type: string
          in: path
          required: true          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetReceiptResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Restaurant:
      type: object
      properties:
        id:
          type: integer      
        userid:
          type: string  
        name:
          type: string
          description: restaurant name  
        phone: 
          type: string
        email: 
          type: string
        note: 
          type: string           
      required:
        - name
  GetRestaurantsListResponse:
      required:
        - restaurants
      properties:
        # The array of movies
        restaurants:
          type: array
          items: 
            type: object
            properties:
              id:
                type: integer
              userid:
                type: string                
              name:
                type: string
  GetRestaurantResponse:
      required:
        - id
        - userid
        - name
      properties:
        id:
          type: integer
        userid:
          type: string
        name: 
          type: string                    
...
  GetCustomersListResponse:
      required:
        - customers
      properties:
        # The array of movies
        customers:
          type: array
          items: 
            type: object
            properties:
              id:
                type: integer
              userid:
                type: string                
              name:
                type: string
  GetCustomerResponse:
      required:
        - id
        - userid
        - name
      properties:
        id:
          type: integer
        userid:
          type: string
        name: 
          type: string
  GetDineTablesListResponse:
      required:
        - dinetables
      properties:
        # The array of movies
        dinetables:
          type: array
          items: 
            type: object
            properties:
              id:
                type: integer
  GetDineTableResponse:
      required:
        - id
      properties:
        id:
          type: integer
  GetWaitsListResponse:
      required:
        - waits
      properties:
        # The array of movies
        waits:
          type: array
          items: 
            type: object
            properties:
              id:
                type: integer
  GetWaitsResponse:
      required:
        - id
      properties:
        id:
          type: integer
...
  GetDishesListResponse:
      required:
        - dishes
      properties:
        # The array of movies
        dishes:
          type: array
          items: 
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
  GetDishResponse:
      required:
        - name
      properties:
        name: 
          type: string
...
  GetDishgroupsListResponse:
      required:
        - dishgroups
      properties:
        # The array of movies
        dishgroups:
          type: array
          items: 
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
  GetDishgroupResponse:
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name: 
          type: string
...
  GetOrdersListResponse:
      required:
        - orders
      properties:
        # The array of movies
        orders:
          type: array
          items: 
            type: object
            properties:
              id:
                type: integer
...
  GetOrderItemsListResponse:
      required:
        - orderitems
      properties:
        # The array of movies
        orderitems:
          type: array
          items: 
            type: object
            properties:
              id:
                type: integer
...
  GetReceiptsListResponse:
      required:
        - receipts
      properties:
        # The array of movies
        receipts:
          type: array
          items: 
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
  GetReceiptResponse:
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name: 
          type: string
...
  GeneralResponse:
      type: object
      properties:
        success:
          type: number
          description: returns 1 if successful
        description:
          type: string
          description: a short comment 
      required:
        - success
        - description
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string